+++ platform/core/src/services/index.js
@@ -1,0 +1,15 @@
+import ServicesManager from './ServicesManager.js';
+import UINotificationService from './UINotificationService';
+import UIModalService from './UIModalService';
+import UIDialogService from './UIDialogService';
+import UIContextMenuService from './UIContextMenuService';
+import UILabellingFlowService from './UILabellingFlowService';
+
+export {
+  UINotificationService,
+  UIModalService,
+  UIDialogService,
+  UIContextMenuService,
+  UILabellingFlowService,
+  ServicesManager,
+};