+++ platform/i18n/src/locales/index.js
@@ -1,3 +1,4 @@
+import ar from './ar/';
 import en_US from './en-US/';
 import es from './es/';
 import ja_JP from './ja-JP/';
@@ -7,6 +8,7 @@ import vi from './vi/';
 import zh from './zh/';
 
 export default {
+  ...ar,
   ...en_US,
   ...es,
   ...ja_JP,