+++ platform/viewer/cypress/support/aliases.js
@@ -11,6 +11,7 @@ export function initCornerstoneToolsAliases() {
   cy.get('.ToolbarRow > :nth-child(10)').as('cineBtn');
   cy.get('.expandableToolMenu').as('moreBtn');
   cy.get('.btn-group > .toolbar-button').as('layoutBtn');
+  cy.get('.viewport-element').as('viewport');
 }
 
 //Creating aliases for Common page elements
@@ -76,4 +77,18 @@ export function initStudyListAliasesOnTablet() {
   cy.get('#start-date').as('studyListStartDate');
   cy.get('#end-date').as('studyListEndDate');
   cy.get('[data-cy="study-list-results"] > tr').as('searchResult');
+}
+
+//Creating aliases for User Preferences modal
+export function initPreferencesModalAliases() {
+  cy.get('.OHIFModal').as('preferencesModal');
+  cy.get('.nav-link')
+    .first()
+    .as('userPreferencesHotkeysTab');
+  cy.get('.nav-link')
+    .last()
+    .as('userPreferencesGeneralTab');
+  cy.get('[data-cy="reset-default-btn"]').as('restoreBtn');
+  cy.get('[data-cy="cancel-btn"]').as('cancelBtn');
+  cy.get('[data-cy="save-btn"]').as('saveBtn');
 }