+++ platform/viewer/public/config/demo.js
@@ -1,5 +1,6 @@
 window.config = {
   routerBasename: '/',
+  whiteLabelling: {},
   extensions: [],
   showStudyList: true,
   servers: {