+++ platform/viewer/src/connectedComponents/Viewer.js
@@ -7,11 +7,11 @@ import OHIF from '@ohif/core';
 import moment from 'moment';
 import ConnectedHeader from './ConnectedHeader.js';
 import ConnectedToolbarRow from './ConnectedToolbarRow.js';
-import ConnectedLabellingOverlay from './ConnectedLabellingOverlay';
 import ConnectedStudyBrowser from './ConnectedStudyBrowser.js';
 import ConnectedViewerMain from './ConnectedViewerMain.js';
 import SidePanel from './../components/SidePanel.js';
 import { extensionManager } from './../App.js';
+import DICOMSR from '../lib/DICOMSR';
 
 // Contexts
 import WhiteLabellingContext from '../context/WhiteLabellingContext.js';
@@ -58,21 +58,31 @@ class Viewer extends Component {
   static propTypes = {
     studies: PropTypes.array,
     studyInstanceUids: PropTypes.array,
+    activeServer: PropTypes.shape({
+      type: PropTypes.string,
+      wadoRoot: PropTypes.string,
+    }),
     onTimepointsUpdated: PropTypes.func,
     onMeasurementsUpdated: PropTypes.func,
     // window.store.getState().viewports.viewportSpecificData
     viewports: PropTypes.object.isRequired,
     // window.store.getState().viewports.activeViewportIndex
     activeViewportIndex: PropTypes.number.isRequired,
+    isStudyLoaded: PropTypes.bool,
   };
 
   constructor(props) {
     super(props);
+
+    const { activeServer } = this.props;
+    const server = Object.assign({}, activeServer);
+
     OHIF.measurements.MeasurementApi.setConfiguration({
       dataExchange: {
-        retrieve: this.retrieveMeasurements,
-        store: this.storeMeasurements,
+        retrieve: DICOMSR.retrieveMeasurements,
+        store: DICOMSR.storeMeasurements,
       },
+      server,
     });
 
     OHIF.measurements.TimepointApi.setConfiguration({
@@ -92,18 +102,6 @@ class Viewer extends Component {
     selectedRightSidePanel: '',
     selectedLeftSidePanel: 'studies', // TODO: Don't hardcode this
     thumbnails: [],
-  };
-
-  retrieveMeasurements = (patientId, timepointIds) => {
-    OHIF.log.info('retrieveMeasurements');
-    // TODO: Retrieve the measurements from the latest available SR
-    return Promise.resolve();
-  };
-
-  storeMeasurements = (measurementData, timepointIds) => {
-    OHIF.log.info('storeMeasurements');
-    // TODO: Store the measurements into a new SR sent to the active server
-    return Promise.resolve();
   };
 
   retrieveTimepoints = filter => {
@@ -172,7 +170,7 @@ class Viewer extends Component {
   };
 
   componentDidMount() {
-    const { studies } = this.props;
+    const { studies, isStudyLoaded } = this.props;
     const { TimepointApi, MeasurementApi } = OHIF.measurements;
     const currentTimepointId = 'TimepointId';
 
@@ -192,8 +190,11 @@ class Viewer extends Component {
       const patientId = studies[0] && studies[0].patientId;
 
       timepointApi.retrieveTimepoints({ patientId });
-      measurementApi.retrieveMeasurements(patientId, [currentTimepointId]);
-
+      if (isStudyLoaded) {
+        this.measurementApi.retrieveMeasurements(patientId, [
+          currentTimepointId,
+        ]);
+      }
       this.setState({
         thumbnails: _mapStudiesToThumbnails(studies),
       });
@@ -201,17 +202,18 @@ class Viewer extends Component {
   }
 
   componentDidUpdate(prevProps) {
-    if (this.props.studies !== prevProps.studies) {
-      const { studies } = this.props;
-      const patientId = studies[0] && studies[0].patientId;
-      const currentTimepointId = this.currentTimepointId;
-
-      this.timepointApi.retrieveTimepoints({ patientId });
-      this.measurementApi.retrieveMeasurements(patientId, [currentTimepointId]);
-
+    const { studies, isStudyLoaded } = this.props;
+    if (studies !== prevProps.studies) {
       this.setState({
         thumbnails: _mapStudiesToThumbnails(studies),
       });
+    }
+    if (isStudyLoaded && isStudyLoaded !== prevProps.isStudyLoaded) {
+      const patientId = studies[0] && studies[0].patientId;
+      const { currentTimepointId } = this;
+
+      this.timepointApi.retrieveTimepoints({ patientId });
+      this.measurementApi.retrieveMeasurements(patientId, [currentTimepointId]);
     }
   }
 
@@ -295,11 +297,11 @@ class Viewer extends Component {
                 activeIndex={this.props.activeViewportIndex}
               />
             ) : (
-              <ConnectedStudyBrowser
-                studies={this.state.thumbnails}
-                studyMetadata={this.props.studies}
-              />
-            )}
+                <ConnectedStudyBrowser
+                  studies={this.state.thumbnails}
+                  studyMetadata={this.props.studies}
+                />
+              )}
           </SidePanel>
 
           {/* MAIN */}
@@ -317,7 +319,6 @@ class Viewer extends Component {
             )}
           </SidePanel>
         </div>
-        <ConnectedLabellingOverlay />
       </>
     );
   }
@@ -336,7 +337,7 @@ export default Viewer;
  * @param {Study[]} studies
  * @param {DisplaySet[]} studies[].displaySets
  */
-const _mapStudiesToThumbnails = function(studies) {
+const _mapStudiesToThumbnails = function (studies) {
   return studies.map(study => {
     const { studyInstanceUid } = study;